Revert to previous version of notes.txt
git checkout can be used in a plethora of ways, most of the time though
it will be used to checkout a branch, or specific commit.
$ git help checkout

To get back a previous version of notes.txt containg information about
initializing the git repo, pass git checkout with the commit hash and
the
specific filename.  Find the correct revision using log, the '--' means
no
more options are passed, so git will interpret the rest as arguments
(which
in this case is filenames).  Then we will rename the

$ git log
$ git log --oneline
$ git checkout 0a29a06 -- notes.txt
$ git commit