Save init and log versions of notes.txt
Similar to last time, we will check out a specific version of the notes
file, however, this time we will move them to their own unique files.
We could have combined this into a single commit because git is able
to keep track of the operations that moved files using its index,
or staging area.
$ git mv notes.txt notes-init.txt
$ git log --oneline
$ git checkout 69183ca -- notes.txt
$ git mv notes.txt notes-log.txt
$ git status
$ git commit